fix: Properly reset original current working directory upon module import
`$currentPath` was originally set inside the `try` block, which has a different scope than the `finally` block. This resulted in `$currentPath` being `$null` when executing in the `finally` block when used in the call to `Set-Location` which effectively uses the default value for the `-Path` parameter to that cmdlet, which is `.`, or the current working directory--or essentially, it doesn't change locations.

Instead, set `$currentPath` outside the `try` block so that when `Set-Location` is executed in the `finally` block, `$currentPath` has the intended value and effects a change to the current working directory.